gs: avoid stupid shifts by casting to uint64_t

+++ sys/src/cmd/gs/src/gdevm48.c
@@ -62,8 +62,8 @@ mem_full_alpha_device("image48", 48, 0, mem_open,
 
 /* Unpack a color into its bytes. */
 #define declare_unpack_color(a, b, c, d, e, f, color)\
-	byte a = (byte)(color >> 40);\
-	byte b = (byte)(color >> 32);\
+	byte a = (byte)((uint64_t)color >> 40);\
+	byte b = (byte)((uint64_t)color >> 32);\
 	byte c = (byte)((uint)color >> 24);\
 	byte d = (byte)((uint)color >> 16);\
 	byte e = (byte)((uint)color >> 8);\
